refactor: archive obsolete scripts, finalize bulk import naming, and update documentation for project structure and script management
