+++ frontend/src/App.tsx
@@ -5,7 +5,7 @@ import ProgramDashboard from './components/features/programs/ProgramDashboard';
 import LedgerPage from './components/features/ledger/LedgerPage';
 import ProgramSettingsPage from './components/features/programs/ProgramSettings';
 import ActualsUploadPage from './components/features/actuals/ActualsUploadPage/index';
-import UploadSessionDetails from './components/features/actuals/UploadSessionDetails';
+import UploadSessionDetails from './components/features/actuals/UploadSessionDetails/index';
 
 // Placeholder components for BOE and Risks & Opportunities
 const BOEPage = () => <div className="p-8"><h1 className="text-2xl font-bold mb-4">BOE Page</h1><p>This is a placeholder for the BOE page.</p></div>;
