+++ frontend/src/components/features/actuals/README.md
@@ -1,21 +1,21 @@
-# Actual Expenses Feature
+# Upload Actuals Feature
 
-This feature handles the upload and processing of monthly actual expense transactions.
+This feature handles the upload and processing of monthly actuals transactions.
 
 ## Components
 
-- **ActualsUploadPage**: Main page for uploading actual expense files
+- **ActualsUploadPage**: Main page for uploading actuals files
 - **TransactionMatchModal**: Modal for reviewing and confirming transaction matches
 - **UploadSessionDetails**: Page for viewing upload session details and results
 
 ## Purpose
 
 This feature allows users to:
-1. Upload monthly actual expense reports from accounting systems (NetSuite, QuickBooks, etc.)
+1. Upload monthly actuals reports from accounting systems (NetSuite, QuickBooks, etc.)
 2. Automatically match transactions to existing ledger entries
 3. Review and confirm matches with confidence scoring
 4. Handle unmatched transactions as new expenses
 
 ## Usage
 
-Users access this feature through the "Actual Expenses" navigation item within a program.
+Users access this feature through the "Upload Actuals" navigation item within a program.
