+++ frontend/src/components/features/actuals/index.ts
@@ -1,3 +1,3 @@
 export { default as ActualsUploadPage } from './ActualsUploadPage';
-export { default as TransactionMatchModal } from './TransactionMatchModal';
-export { default as UploadSessionDetails } from './UploadSessionDetails';
+export { default as TransactionMatchModal } from './TransactionMatchModal/index';
+export { default as UploadSessionDetails } from './UploadSessionDetails/index';
