+++ frontend/src/components/features/ledger/BulkImport/README.md
@@ -7,9 +7,9 @@ This feature handles the initial setup of program ledger entries through bulk im
 This feature allows users to:
 1. Import initial program budget and planned expenses
 2. Set up baseline spending categories
-3. Bulk add multiple ledger entries at once
+3. Bulk import multiple ledger entries at once
 4. Use templates for consistent data entry
 
 ## Usage
 
-Users access this feature through the "Bulk Add" button on the Ledger page.
+Users access this feature through the "Bulk Import" button on the Ledger page.
