+++ frontend/src/components/features/ledger/index.ts
@@ -1,3 +1,3 @@
-// export { default as LedgerPage } from '../LedgerPage';
-// export { default as LedgerTable } from '../LedgerTable';
-export {};
+export { default as LedgerPage } from './LedgerPage';
+export { default as LedgerTable } from './LedgerTable';
+export { default as BulkImportModal } from './BulkImport/BulkImportModal';
